+++ nzbhydra/search.py
@@ -4,6 +4,7 @@ from __future__ import print_function
 from __future__ import unicode_literals
 
 import re
+from datetime import timedelta
 from itertools import groupby
 
 from builtins import int
@@ -17,7 +18,7 @@ import copy
 import logging
 import arrow
 from requests_futures.sessions import FuturesSession
-from nzbhydra.database import IndexerStatus, Search, db
+from nzbhydra.database import IndexerStatus, Search, db, IndexerApiAccess
 from nzbhydra import config, indexers, infos
 
 categories = {'All': {"pretty": "All", "index": 0},
@@ -122,6 +123,23 @@ def pick_indexers(search_request, internal=True):
                 continue
         except IndexerStatus.DoesNotExist:
             pass
+        
+        if p.settings.hitLimit > 0:
+            if p.settings.hitLimitResetTime:
+                hitLimitResetTime = arrow.get(p.settings.hitLimitResetTime)
+                comparisonTime = arrow.now().replace(hour=hitLimitResetTime.hour, minute=hitLimitResetTime.minute, second=0)
+                if comparisonTime > arrow.now():
+                    comparisonTime = arrow.get(comparisonTime.datetime - timedelta(days=1)) #Arrow is too dumb to properly subtract 1 day (throws an error on every first of the month)
+            else:
+                comparisonTime = arrow.now().replace(hour=0, minute=0, second=0)
+            
+            apiHits = IndexerApiAccess().select().where((IndexerApiAccess.indexer == p.indexer) & (IndexerApiAccess.time > comparisonTime)).count()
+            if apiHits > p.settings.hitLimit:
+                logger.info("Did not pick %s because its API hit limit of %d was reached" % (p, p.settings.hitLimit))
+                continue
+            else:
+                logger.debug("%s has had %d of a maximum of %d API hits since %02d:%02d" % (p, apiHits, p.settings.hitLimit, comparisonTime.hour, comparisonTime.minute))
+        
 
         if (query_supplied or search_request.identifier_key is not None)and not p.supports_queries:
             logger.debug("Did not pick %s because a query was supplied but the indexer does not support queries" % p)