+++ ui-src/js/config-fields-service.js
@@ -130,6 +130,32 @@ function ConfigFields() {
                 }
             }
         ]);
+        
+        if (testtype == "newznab") {
+            fieldset.push(
+                {
+                    key: 'hitLimit',
+                    type: 'horizontalInput',
+                    hideExpression: '!model.enabled',
+                    templateOptions: {
+                        type: 'number',
+                        label: 'API hit limit',
+                        help: 'Maximum number of API hits since "API hit reset time"'
+                    }
+                }
+            );
+            fieldset.push(
+                {
+                    key: 'hitLimitResetTime',
+                    type: 'timeOfDay',
+                    hideExpression: '!model.enabled || !model.hitLimit',
+                    templateOptions: {
+                        type: 'time',
+                        label: 'API hit reset time',
+                        help: 'Local time at which the API hit counter is reset'
+                    }
+                })
+        }
 
 
         if (showpreselect) {
@@ -1179,6 +1205,8 @@ function ConfigFields() {
                             enabled: true,
                             host: null,
                             apikey: null,
+                            hitLimit: null,
+                            hitLimitResetTime: new Date(0),
                             timeout: null,
                             name: null,
                             showOnSearch: true,